extract link-container to own file
