+++ gatsby-theme-landing-page/src/components/benefits.js
@@ -2,7 +2,8 @@ import * as React from "react";
 import { GatsbyImage, getImage } from "gatsby-plugin-image";
 import * as styles from "./benefits.module.css";
 import MarkdownText from "./markdown-text";
-import Link, { LinkContainer } from "./link";
+import Link from "./link";
+import LinkContainer from "./link-container";
 
 export default function Benefits({ heading, secondaryHeading, content }) {
   return (
