+++ gatsby-theme-landing-page/src/components/link-container.module.css
@@ -1,0 +1,11 @@
+.linkContainer {
+  display: flex;
+  flex-direction: column;
+  gap: var(--space-4);
+}
+
+@media screen and (min-width: 480px) {
+  .linkContainer {
+    flex-direction: row;
+  }
+}
